cmd_gen script for parboil + memory divergence calc in get_primary_gpu.sh

+++ src/get_primary_gpu.sh
@@ -45,25 +45,33 @@ if [ $DEBUG ]; then
     echo $execstr
 fi
 
-(nvprof -u ms --system-profiling on $execstr > prog.out) 2> tmp
+if [ ${metric} = "time" ] || [ ${metric} = "pwr" ]; then 
 
-if [ "$kernel" = "none" ]; then 
-	time=`cat tmp | grep "Time(%)" -m 1 -A 2 2>&1 | tail -1 | awk '{print $2/($1/100)}'`
-else 
-	time=`cat tmp | grep  ${kernel} | awk '{printf $2 " " $4}'`
-fi
-time=`echo $time | awk '{printf "%3.2f", $1}'`
+	(nvprof -u ms --system-profiling on $execstr > prog.out) 2> tmp
 
-pwr=`cat tmp  | grep "Power" | awk '{print $4}'`
-pwr=`echo $pwr | awk '{printf "%3.2f", $1/1000}'`
-
-if [ ${metric} = "time" ]; then 
+	if [ "$kernel" = "none" ]; then 
+		time=`cat tmp | grep "Time(%)" -m 1 -A 2 2>&1 | tail -1 | awk '{print $2/($1/100)}'`
+	else 
+		time=`cat tmp | grep  ${kernel} | awk '{printf $2 " " $4}'`
+	fi
+	time=`echo $time | awk '{printf "%3.2f", $1}'`
+	
+	pwr=`cat tmp  | grep "Power" | awk '{print $4}'`
+	pwr=`echo $pwr | awk '{printf "%3.2f", $1/1000}'`
+	
+	if [ ${metric} = "time" ]; then 
     echo $time 
-fi
-
-if [ ${metric} = "pwr" ]; then 
+	fi
+	
+	if [ ${metric} = "pwr" ]; then 
     echo $pwr 
+	fi
+  # clean up 
+  rm -rf tmp prog.out
 fi
 
-# clean up 
-rm -rf tmp prog.out
+
+if [ ${metric} = "memdiv" ]; then 
+  ctrs=`nvprof --metrics gld_transactions_per_request,gst_transactions_per_request $execstr 2>&1 | grep "transactions_per_request" | awk '{print $NF}'`
+	echo $ctrs 
+fi
